+++ lib/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @flare-ts/lib
 
+## 0.3.0
+
 ## 0.2.0
 
 ### Minor Changes